Removed Comment, dependent from CSV Main function on top now ValueSaver interface renamed to DataProcessor summarizer.go.Storage type renamed to Result Summarizer processors structures( avg.go, count-unique.go, sum.go ) renamed File reader and StdIn functions moved to input.go Flag package used to determine using of «-stdin» command line parameter to change source to stdin Itoa and FormatInt used instead of Sprintf to convert int/int64 type numbers to string ProcessRow function now part of Storage type, so this func became shorter and more flexible
