+++ summarizer/interface.go
@@ -1,0 +1,6 @@
+package summarizer
+
+type DataProcessor interface {
+	AddValue(string) error
+	Result() string
+}
